Amost Finished 5.2.3. Did not deal with invalid products

+++ app/controllers/products_controller.rb
@@ -2,6 +2,10 @@ class ProductsController < ApplicationController
   before_action :authenticate_account!
   before_action :set_product, only: [:show, :edit, :update, :destroy]
 
+  def pundit_user
+    current_account
+  end
+
   # GET /products
   # GET /products.json
   def index
@@ -10,28 +14,35 @@ class ProductsController < ApplicationController
       @products = @seller.products
     else
       @products = Product.all
-    end   
+    end
+
+    authorize Product 
+    @products = policy_scope(Product)
   end
 
   # GET /products/1
   # GET /products/1.json
   def show
+    authorize @product 
     @orders = @product.orders
   end
 
   # GET /products/new
   def new
     @product = Product.new
+    authorize @product 
   end
 
   # GET /products/1/edit
   def edit
+    authorize @product 
   end
 
   # POST /products
   # POST /products.json
   def create
     @product = Product.new(product_params)
+    authorize @product 
 
     if current_account && current_account.accountable_type == "Seller"
         @product.seller = current_account.accountable
@@ -51,6 +62,7 @@ class ProductsController < ApplicationController
   # PATCH/PUT /products/1
   # PATCH/PUT /products/1.json
   def update
+    authorize @product 
     respond_to do |format|
       if @product.update(product_params)
         format.html { redirect_to @product, notice: 'Product was successfully updated.' }
@@ -69,6 +81,7 @@ class ProductsController < ApplicationController
   # DELETE /products/1
   # DELETE /products/1.json
   def destroy
+    authorize @product 
     respond_to do |format|
       if @product.destroy
         format.html { redirect_to products_url, notice: 'Product was successfully destroyed.' }
